Introduce `esbonio-eglot-ensure`
This appears to be the cleanest way to ensure that the user can both
defer the loading of `esbonio.el` while still ensuring that esbonio
does load itself before eglot attempts to connect to the server.

This commit also tries to ensure that when the package is installed it
compiles cleanly. However, I still get this warning, despite clearly
having a `lexical-binding: t` declaration at the top of esbonio.el

  Warning (bytecomp): file has no ‘lexical-binding’ directive on its first line